Adds support for Authenticated Tor Hidden Services
The ATHS requires connecting clients to provide an auth cookie,
specified as HidServAuth in the client torrc, in order to communicate
with the hidden service. Specifying a hidden service authenticated in
this role now requires an additional key in the service config dict.

The new functionality is off by default. The logic displaying hostnames
still displays the HidServAuth info, but the formatting could be better.