Special handling for HidServAuth config line
The HidServAuth line can be repeated, and must be, so assume that vars
for this config option are a list, and iterate over them. Haven't
searched the tor manpage for other listable vars; in the future it may
be convenient to store a role default var with a list of listvars,
and update the template to reference that.