support for onion v3 secret keys
Onion v3 private keys are stored in the binary form on the file
hs_ed25519_secret_key, which makes it impractical to write in a
variable. The solution was to define it in base64.

It should be noted that this will break if the file
'private_key_encoded' is opened in a common text editor and saved,
as some modern editors add a trailing '\n' to the file automatically.
As that would change the value stored in key file.